Generalize ShareDBStringBinding React component
