+++ examples/textarea-routing/src/TextPad.css
@@ -1,4 +1,4 @@
-.TextPad {
+.text-pad {
   font-size: 2em;
   height: 100%;
   width: 100%;
